Added restrictions for Test user

+++ src/modules/user/user.controller.ts
@@ -2,7 +2,6 @@ import mongoose from 'mongoose';
 import * as bcrypt from 'bcrypt';
 import {JwtService} from '@nestjs/jwt';
 import {UserService} from './user.service';
-import {IResponse} from '@/types/app.types';
 import {AuthSuccessMessages} from '@messages/auth';
 import {UserSuccessMessages} from '@messages/user';
 import {CardService} from '@cardModule/card.service';
@@ -12,11 +11,12 @@ import {AuthGuard} from '@authModule/guards/auth.guard';
 import {FileInterceptor} from '@nestjs/platform-express';
 import {ICardResponse} from '@cardModule/types/card.types';
 import {CommonService} from '@commonModule/common.service';
+import {ICustomRequest, IResponse} from '@/types/app.types';
 import {UpdateUserProfileDto} from './dto/update-user-profile.dto';
 import {IUpdateUserProfile, IUserResponse} from './types/user.types';
 import {UpdateUserSecurityDto} from './dto/update-user-security.dto';
 import {TransactionService} from '@transactionModule/transaction.service';
-import {Controller, Get, Body, Put, Param, Delete, UseInterceptors, UploadedFile, HttpStatus, HttpException, UseGuards} from '@nestjs/common';
+import {Controller, Get, Body, Put, Param, Delete, UseInterceptors, UploadedFile, HttpStatus, HttpException, UseGuards, Request} from '@nestjs/common';
 
 @Controller('user')
 @UseGuards(AuthGuard)
@@ -32,8 +32,8 @@ export class UserController {
   ) {}
 
   @Get(':id')
-  async findOne(@Param('id') id: string): Promise<IResponse<IUserResponse>> {
-    const response = await this.userService.findOne(id);
+  async findOne(@Request() req: ICustomRequest, @Param('id') id: string): Promise<IResponse<IUserResponse>> {
+    const response = await this.userService.findOne(id, req.role);
     return ({
       data: response,
       statusCode: HttpStatus.OK,
@@ -42,23 +42,27 @@ export class UserController {
   }
 
   @Delete(':id')
-  async removeOne(@Param('id') id: string): Promise<IResponse<undefined>> {
-    const user = await this.commonService.findOneUserAPI('_id', id);
-    const cards = await this.cardService.findMany({ownerId: new mongoose.Types.ObjectId(id)});
-    cards.map(async (el: ICardResponse) => { // delete all cards transactions
-      await this.transactionService.removeMany(el._id.toString());
-    });
-    await this.cardService.removeMany(id); // delete all users cards
-    await this.imageService.removeOne(user.imageId); // delete image(avatar)
-    const response = await this.userService.removeOne(id);
-    return ({
-      message: response,
-      statusCode: HttpStatus.OK,
-    });
+  async removeOne(@Request() req: ICustomRequest, @Param('id') id: string): Promise<IResponse<undefined>> {
+    if(req.role === 'Admin' || (req._id === id && req.role !== 'Test')) {
+      const user = await this.commonService.findOneUserAPI('_id', id);
+      const cards = await this.cardService.findMany({ownerId: new mongoose.Types.ObjectId(id)});
+      cards.cards.map(async (el: ICardResponse) => { // delete all cards transactions
+        await this.transactionService.removeMany(el._id.toString());
+      });
+      await this.cardService.removeMany(id); // delete all users cards
+      await this.imageService.removeOne(user.imageId); // delete image(avatar)
+      const response = await this.userService.removeOne(id);
+      return ({
+        message: response,
+        statusCode: HttpStatus.OK,
+      });
+    }
+    throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
   }
 
   @Delete('avatar/:id')
-  async removeAvatar(@Param('id') id: string): Promise<IResponse<undefined>> {
+  async removeAvatar(@Request() req: ICustomRequest, @Param('id') id: string): Promise<IResponse<undefined>> {
+    if(req._id !== id) throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
     const user = await this.commonService.findOneUserAPI('_id', id);
     if(user.imageId) {
       await this.imageService.removeOne(user.imageId);
@@ -71,7 +75,8 @@ export class UserController {
   }
 
   @Put('security/:id')
-  async updateSecurity(@Param('id') id: string, @Body() updateUserSecurityDto: UpdateUserSecurityDto): Promise<IResponse<undefined>> {
+  async updateSecurity(@Request() req: ICustomRequest, @Param('id') id: string, @Body() updateUserSecurityDto: UpdateUserSecurityDto): Promise<IResponse<undefined>> {
+    if(req._id !== id || req.role === 'Test') throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
     const user = await this.commonService.findOneUserAPI('_id', id, true);
     const isPassValid = bcrypt.compareSync(updateUserSecurityDto.oldPassword, user.password);
     if(!isPassValid) throw new HttpException('Wrong old password', HttpStatus.BAD_REQUEST);
@@ -84,8 +89,9 @@ export class UserController {
   }
 
   @Put('profile/:id')
-  @UseInterceptors(FileInterceptor('avatar'))
-  async updateProfile(@Param('id') id: string, @UploadedFile() file: Express.Multer.File, @Body() updateUserProfileDto: UpdateUserProfileDto): Promise<IResponse<undefined>> {
+  @UseInterceptors(FileInterceptor('file'))
+  async updateProfile(@Request() req: ICustomRequest, @Param('id') id: string, @UploadedFile() file: Express.Multer.File, @Body() updateUserProfileDto: UpdateUserProfileDto): Promise<IResponse<undefined>> {
+    if(req._id !== id) throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
     const updateUserProfile: IUpdateUserProfile = {};
     const user = await this.commonService.findOneUserAPI('_id', id);
     if(file) {
@@ -113,9 +119,9 @@ export class UserController {
       await this.commonService.findOneCurrency(updateUserProfileDto.currency);
       updateUserProfile.currency = updateUserProfileDto.currency;
     }
-    if(updateUserProfileDto.email) updateUserProfile.email = updateUserProfileDto.email;
     if(updateUserProfileDto.lastName) updateUserProfile.lastName = updateUserProfileDto.lastName;
     if(updateUserProfileDto.firstName) updateUserProfile.firstName = updateUserProfileDto.firstName;
+    if(updateUserProfileDto.email && req.role !== 'Test') updateUserProfile.email = updateUserProfileDto.email;
     await this.userService.updateProfile(id, updateUserProfile);
     if(updateUserProfile.email) {
       const code = this.jwtService.sign({email: updateUserProfile.email, id, password: user.password}, {expiresIn: process.env.EMAIL_CODE_EXPIRES_IN});
